second commit

+++ app/src/main/java/com/example/appple/appcalhw/events.java
@@ -4,28 +4,28 @@ package com.example.appple.appcalhw;
  * Created by appple on 3/11/16.
  */
 public class events {
-    private int id;
+    private Long id;
     private String eventname;
     private java.util.Date date;
 
     public events() {
     }
 
-    public events(int id) {
+    public events(long id) {
         this.id = id;
     }
 
-    public events(int id, String eventname, java.util.Date date) {
+    public events(long id, String eventname, java.util.Date date) {
         this.id = id;
         this.eventname = eventname;
         this.date = date;
     }
 
-    public int getId() {
+    public long getId() {
         return this.id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
